+++ stisim/interventions.py
@@ -10,5 +10,9 @@ __all__ = ['Intervention']
 
 class Intervention(ss.Module):
 
-    def apply(self):
+
+    def __call__(self, *args, **kwargs):
+        return self.apply(*args, **kwargs)
+
+    def apply(self, sim, *args, **kwargs):
         raise NotImplementedError
